+++ src/OpenTelemetry.Resources.Azure/CHANGELOG.md
@@ -11,6 +11,9 @@
 * Fix `os.type` not being normalized to lowercase for Azure VMs.
   ([#4775](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4775))
 
+* Add event source to log exceptions during resource detection.
+  ([#4775](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4775))
+
 ## 1.16.0-beta.1
 
 Released 2026-Jul-10